refactor wing portion of the function and rename file
